+++ application/controllers/Flights.php
@@ -110,8 +110,6 @@ class Flights extends Application
             
             $this->render();
         }
-
-
     }
 
     public function add()
@@ -120,10 +118,13 @@ class Flights extends Application
         $this->load->library('form_validation');
         $this->form_validation->set_rules($this->flightModel->formRules());
         $data = $this->input->post();
+        $add = array("id" => null);
         // validate away
         if ($this->form_validation->run())
         {
-            unset($data['id']);
+//            unset($data['id']);
+            //Needed to set this array merge cause PHP was complaining about not having an id
+            $data = array_merge($add, $data);
             $flight = new Flight($data);
             $flight_id = $this->flightModel->saveFlight($flight);
 
@@ -154,5 +155,93 @@ class Flights extends Application
             redirect('/flights');
         }
     }
+
+//    // Initiate adding a new flight
+//    public function add()
+//    {
+//        $this->load->model('flightModel');
+//        $flightModel = $this->flightModel->create();
+//        $this->session->set_userdata('flightModel', $flightModel);
+//        $this->showit();
+//    }
+
+    // submit
+    public function submit()
+    {
+        // FORM VALIDATION HAS NOT BEEN IMPLEMENTED
+        // ...
+
+        // retrieve & update data transfer buffer
+        $this->load->model('flightModel');
+        $task = (array)$this->session->userdata('flightModel');
+        $task = array_merge($task, $this->input->post());
+        $task = (object)$task;  // convert back to object
+        //var_dump($task);
+
+        $this->flightModel->add($task);
+        redirect('/flights');
+
+    }
+
+//    // Render the current DTO
+//    private function showit()
+//    {
+//        $this->load->helper('form');
+//        $task = $this->session->userdata('flightModel');
+////        var_dump($task);
+////        $this->data['id'] = $task->id;
+////
+////        // if no errors, pass an empty message
+////        if ( ! isset($this->data['error']))
+////            $this->data['error'] = '';
+//
+////        {fid}
+////        {fplaneid}
+////        {fdepart}
+////        {fdepart_airport}
+////        {fdepart_time}
+////        {farrival}
+////        {farrival_airport}
+////        {farrival_time}
+////        {zsubmit}
+//
+//        //THE BELOW SHOULD IDEALLY BE CALLED FROM APP MODEL.
+//        //I DIDN'T MANAGE TO MAKE APP MODEL WORK :(
+//        // Planes
+//        $planeid = [
+//        1 => 'Grand Caravan Ex',
+//        2 => 'PC-12 NG',
+//        3 => 'Phenom 100'
+//        ];
+//
+//        // Airport Codes
+//        $airportcodes = [
+//        1	 => 'YVE',
+//        2	 => 'YGE',
+//        3	 => 'ZMH',
+//        4	 => 'YYJ',
+//        5    => 'ZMH'
+//        ];
+//
+//        // if no errors, pass an empty message
+//        if ( ! isset($this->data['error']))
+//            $this->data['error'] = '';
+//
+//        $fields = array(
+//            'fid'      => form_label('Flight ID') . form_input('id', $task->id),
+//            'fplaneid'  => form_label('Plane ID') . form_dropdown('plane', $planeid, $task->plane),
+//            'fdepart'      => form_label('Departure Location') . form_dropdown('depart', $airportcodes, $task->depart),
+//            'fdepart_airport'      => form_label('Departure Airport') .form_input('depart_airport', $task->depart_airport),
+//            'fdepart_time'      => form_label('Departure Time') . form_input('depart_time', $task->depart_time),
+//            'farrival'      => form_label('Arrival Location') . form_dropdown('arrival', $airportcodes, $task->arrival),
+//            'farrival_airport'      => form_label('Arrival Airport') . form_input('arrival_airport', $task->arrival_airport),
+//            'farrival_time'      => form_label('Arrival Time') . form_input('arrival_time', $task->arrival_time),
+//            'zsubmit'    => form_submit('submit', 'Create Flight Details'),
+//        );
+//        $this->data = array_merge($this->data, $fields);
+//
+//        $this->data['pagebody'] = 'itemedit';
+//        $this->render();
+//    }
 
 }